refactor: quantity/stock 타입을 Long으로 통일

+++ apps/commerce-api/src/main/kotlin/com/loopers/application/order/OrderCommand.kt
@@ -4,4 +4,4 @@ import java.math.BigDecimal
 
 data class OrderCommand(val orderItems: List<OrderItemCommand>)
 
-data class OrderItemCommand(val productId: Long, val quantity: Int, val productPrice: BigDecimal)
+data class OrderItemCommand(val productId: Long, val quantity: Long, val productPrice: BigDecimal)
